examples/plugins/ragbrain-example-provider/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stache-example-provider"
version = "0.1.0"
description = "Example external provider for Stache"
requires-python = ">=3.10"
dependencies = [
    "stache-ai>=0.1.0",
]

[project.entry-points."stache.vectordb"]
example = "stache_example.provider:ExampleVectorDBProvider"
